Some light refactoring
